Removed all warnings in generalized_hough
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@20774 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 42474fd1af0bc8d4dc8b83e22def501f8a533e8e

+++ stacks/visual_feature_detectors/outlet_detection/include/outlet_detection/one_way_descriptor.h
@@ -32,7 +32,7 @@ inline int round(float value)
 
 inline CvRect resize_rect(CvRect rect, float alpha)
 {
-	return cvRect(rect.x + round(0.5*(1 - alpha)*rect.width), rect.y + round(0.5*(1 - alpha)*rect.height), 
+	return cvRect(rect.x + round((float)(0.5*(1 - alpha)*rect.width)), rect.y + round((float)(0.5*(1 - alpha)*rect.height)), 
 				  round(rect.width*alpha), round(rect.height*alpha));
 }
